+++ index.js
@@ -6,6 +6,9 @@ const path = require('node:path');
 const { Client, Collection, Events, GatewayIntentBits, AttachmentBuilder } = require('discord.js');
 const { OpenAI }  = require('openai');
 
+// Import the moderation function from the moderation.js file
+const moderateMessage = require('./moderation');
+
 /// List of words to check for in messages
 const naughtyWords = ['duck', 'spit']
 /// List of users who have been warned about using banned words or flagged content
@@ -74,7 +77,11 @@ client.on(Events.InteractionCreate, async interaction => {
             await interaction.reply({ content: 'There was an error while executing this command!', ephemeral: true });
         }
     }
+
+    // utilise moderation function
+    await moderateMessage(message, naughtyWords, banList);
 });
+
 
 
 //---------------------------------------------------------GENERATE IMAGE FUNCTION---------------------------------------------------------------------
@@ -125,15 +132,25 @@ async function handleMessage(message) {
         // Return the image url from the generateImage function
         const imageUrl = await generateImage(prompt);
 
-        // If the image url is not null, send the image in a DM to the user
-        if (imageUrl) {
-            const imageAttachment = new AttachmentBuilder(imageUrl, { name: 'generated-image.png' });
-            // message.channel.send({ files: [imageAttachment] });
-            message.author.send({ content: `Here's your generated image based on the description: "${prompt}"`,
-            files: [imageAttachment]});
-        } else {
-            message.channel.send('Sorry, I was unable to generate an image.');
+       // If the image url is not null, send the image in a DM to the user
+    if (imageUrl) {
+        const imageAttachment = new AttachmentBuilder(imageUrl, { name: 'generated-image.png' });
+
+        // Ensure we're correctly opening a DM channel and sending the message there
+        try {
+            const dmChannel = await message.author.createDM(); // This opens a DM channel
+            await dmChannel.send({ // Use the send method on the DM channel
+                content: `Here's your generated image based on the description: "${prompt}"`,
+                files: [imageAttachment]
+            });
+        } catch (error) {
+            console.error('Failed to send DM:', error);
+            // Optionally, handle failure (e.g., inform the user in the original channel, log the error, etc.)
         }
+    } else {
+    await message.channel.send('Sorry, I was unable to generate an image.');
+    }   
+
     } else {
         // Initialise a conversation with system message
         let conversationLog = [{ role: 'system', content: "You are a friendly chatbot that speaks only in limericks." }];
@@ -169,51 +186,7 @@ async function handleMessage(message) {
             return;
         }
     }
-
-// ---------------------------------------------------------MODERATION-----------------------------------------------------------------------------
-       
-
-    // Convert message content to lower case to make the check case-insensitive
-    const messageContentLowerCase = message.content.toLowerCase();
-
-    // Check if the message contains any of the naughty words
-    const containsNaughtyWord = naughtyWords.some(keyword => messageContentLowerCase.includes(keyword));
-
-    // Use OpenAI automoderation to check for inappropriate content
-    const moderation = await openai.moderations.create({ input: message.content });
-    const flaggedByOpenAI = moderation.results[0].flagged;
-
-    // If the message contains a naughty word or is flagged by OpenAI, add the user to the ban list and send a DM warning
-    if (containsNaughtyWord || flaggedByOpenAI) {
-        // Add user to ban list if not already added
-        if (!banList.includes(message.author.id)) {
-            banList.push(message.author.id);
-            try {
-                await message.author.send('Please refrain from using naughty or inappropriate words.');
-            } catch (error) {
-                console.error(`Could not send DM to ${message.author.tag}.`, error);
-            }
-        }
-
-        // If the user is in the ban list and still uses a naughty word or flagged content
-        if (banList.includes(message.author.id)) {
-            // Delete the message
-            try {
-                await message.delete();
-            } catch (error) {
-                console.error(`Could not delete message from ${message.author.tag}.`, error);
-            }
-
-            // Ban the user from the server
-            try {
-                await message.guild.members.ban(message.author.id, { reason: 'Violating the ban list rules after being warned.' });
-                console.log(`${message.author.tag} has been banned from the server.`);
-            } catch (banError) {
-                console.error(`Could not ban ${message.author.tag}:`, banError);
-            }    
-        }
-    }
-};
+}
 
 client.on('messageCreate', handleMessage);
