+++ package.json
@@ -5,8 +5,10 @@
   "main": "index.js",
   "scripts": {
     "dev": "node index.js",
+    "test": "node --test",
     "discord_test": "node test/discord.test.js",
-    "openai_test": "node test/openai.test.js"
+    "openai_test": "node test/openai.test.js",
+    "env_test": "node test/env.test.js"
   },
   "keywords": [],
   "author": "",
