+++ test/openai.test.js
@@ -1,6 +1,5 @@
 // Import necessary modules and files
 const { handleMessage } = require('../index.js'); // handleMessage function is exported from index.js
-const fs = require('fs'); // Require file system for .env testing
 const { test } = require('node:test');
 const assert = require('assert');
 
@@ -26,46 +25,40 @@ test("OpenAI responding", () => {
     }
 });
 
-// check that my bot securely loads API keys from the .env file, confirming that no sensitive information is hard-coded
-test(".env exists", () => {
-    const envFilePath = '.env';
 
-    try {
-        fs.accessSync(envFilePath, fs.constants.F_OK);
-        console.info("Pass: .env file exists.");
-    } catch (error) {
-        console.error("Fail: .env file does not exist.");
-    }
-});
-
-// check that the .env file is in the gitignore
-test(".env inclusion in gitignore", () => {
-    const gitignoreFilePath = '.gitignore';
-    const gitignoreContent = fs.readFileSync(gitignoreFilePath, 'utf8');
-    assert(gitignoreContent.includes('.env'), "Fail: .env file is not included in .gitignore.")
-})
-
-// check that relevant variables are imported from .env
-test("Import variables from .env", () => {
-    try {
-        // Define the expected environment variables
-        const expectedEnvVariables = {
-            TOKEN: process.env.TOKEN,
-            CHANNEL_ID: process.env.CHANNEL_ID,
-            OPENAI_API_KEY: process.env.OPENAI_API_KEY,
-            CLIENT_ID: process.env.CLIENT_ID
-        };
-
-        // Check if variables are defined and have expected values in index.js
-        for (const [key, value] of Object.entries(expectedEnvVariables)) {
-            assert.ok(value, `Fail: ${key} is not correctly imported in index.js.`);
+// test Dall-E image generation with async/await
+test("Dall-E image generation", async () => {
+    const testMsg = {
+        content: "!image a cat",
+        author: {
+            bot: false // Ensure the message is treated as coming from a real user
+        },
+        channel: {
+            send: (message) => {
+                // This mock function intercepts messages sent by your bot to the channel
+                console.log('Bot replied with:', message);
+                // Basic validation to check if the bot's reply seems like an image URL
+                if (typeof message === 'string' && message.startsWith('http')) {
+                    console.log('Passed: Bot replied with an image URL!');
+                    return true; // Indicate success for the test
+                } else {
+                    throw new Error('Bot did not reply with an image URL.');
+                }
+            }
+        },
+        reply: (response) => {
+            // Alternative reply handler, depending on your bot's implementation
+            console.log('Bot replied with:', response);
         }
+    };
 
-        setTimeout(() => {
-            console.log("Pass: Environment variables are correctly imported in index.js.");
-        }, 0); // Delaying the logging to ensure the message logs after the test result
+    try {
+        // Wait for handleMessage to complete its operation
+        await handleMessage(testMsg);
+        // If handleMessage completes without errors, and send function logic passes, test is considered successful
+        console.log("Test passed: Image generation and response successful.");
     } catch (error) {
-        console.error(error.message); // Log any assertion errors
-        throw error; // Throw the error explicitly to ensure the test fails
+        // Catch errors both from handleMessage and the send function's thrown error
+        console.error(`Test failed: ${error.message}`);
     }
 });
